% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_historical_stacked.R
\name{average_stacked_forecasts}
\alias{average_stacked_forecasts}
\title{Average ensemble from stacked historical noaa forecasts}
\usage{
average_stacked_forecasts(
  forecast_dates,
  site,
  noaa_hour,
  noaa_stacked_directory,
  output_file
)
}
\arguments{
\item{output_file}{}
}
\value{

}
\description{
Average ensemble from stacked historical noaa forecasts
}
