% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DODGrab.R
\name{DODSGrab}
\alias{DODSGrab}
\title{Title}
\usage{
DODSGrab(
  model.url,
  model.run,
  variables,
  time,
  lon,
  lat,
  levels = NULL,
  ensembles = NULL,
  display.url = TRUE,
  verbose = FALSE,
  request.sleep = 1
)
}
\arguments{
\item{request.sleep}{}
}
\value{

}
\description{
Title
}
