% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_downscale.R
\name{downscale_repeat_6hr_to_hrly}
\alias{downscale_repeat_6hr_to_hrly}
\title{Downscale repeat to hourly}
\usage{
downscale_repeat_6hr_to_hrly(df, varName, hr = 1)
}
\arguments{
\item{df, }{dataframe of data to be downscaled (Longwave)}
}
\value{
A dataframe of downscaled data
}
\description{
Downscale repeat to hourly
}
\author{
Laura Puckett
}
