% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_downscale.R
\name{downscale_solar_geom}
\alias{downscale_solar_geom}
\title{Calculate potential shortwave radiation}
\usage{
downscale_solar_geom(doy, lon, lat)
}
\arguments{
\item{doy, }{day of year in decimal}

\item{lon, }{longitude}

\item{lat, }{latitude}
}
\value{
vector of potential shortwave radiation for each doy

\code{numeric(1)}
}
\description{
Calculate potential shortwave radiation
}
\author{
Quinn Thomas
}
