% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_downscale.R
\name{downscale_spline_to_hrly}
\alias{downscale_spline_to_hrly}
\title{Downscale spline to hourly}
\usage{
downscale_spline_to_hrly(df, VarNames, hr = 1)
}
\arguments{
\item{df, }{dataframe of data to be downscales}

\item{VarNames, }{names of vars that are state variables}
}
\value{
A dataframe of downscaled state variables
}
\description{
Downscale spline to hourly
}
\author{
Laura Puckett
}
