% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_cfs_grid_download.R
\name{noaa_cfs_grid_download}
\alias{noaa_cfs_grid_download}
\title{Download gridded forecast in the box bounded by the latitude and longitude list}
\usage{
noaa_cfs_grid_download(
  lat_list,
  lon_list,
  forecast_time,
  forecast_date,
  model_name_raw,
  output_directory,
  grid_name
)
}
\arguments{
\item{output_directory}{}
}
\value{

}
\description{
Download gridded forecast in the box bounded by the latitude and longitude list
}
