% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_cfs_grid_process_downscale.R
\name{noaa_cfs_grid_process_downscale}
\alias{noaa_cfs_grid_process_downscale}
\title{Extract and downscale points from a NOAA CFS Forecast}
\usage{
noaa_cfs_grid_process_downscale(
  lat_list,
  lon_list,
  site_list,
  downscale,
  debias = FALSE,
  overwrite,
  model_name,
  model_name_ds,
  model_name_ds_debias,
  model_name_raw,
  debias_coefficients = NULL,
  num_cores = 1,
  output_directory,
  reprocess = FALSE,
  grid_name
)
}
\arguments{
\item{lat_list}{Vector or range of latitudes to be downloaded (see details).}

\item{lon_list}{Vector or range of longitudes to be downloaded (see details).}

\item{site_list}{Vector of site codes, e.g. "NOAA", used in directory and file name generation.}

\item{downscale}{Logical specifying whether to downscale from 6-hr to 1-hr data.}

\item{debias}{Logical specifying whether to downscale from 6-hr to 1-hr data.}

\item{overwrite}{Logical stating whether to overwrite any existing output files}

\item{model_name}{Directory name for the 6-hr forecast, this will be used in directory and file
name generation.}

\item{model_name_ds}{Directory name for downscaled 1-hr forecasts, this will be used in
directory and file name generation.}

\item{model_name_ds_debias}{Directory name for downscaled and diabiased forecasts, this will be
used in directory and file name generation.}

\item{debias_coefficients}{Data frame containing debasing parameter values.}

\item{num_cores}{DEPRECATED?????}

\item{output_directory}{Path: directory where the model output will be saved.}

\item{reprocess}{Logical: if true re-extract dates that were previously processed.}

\item{grid_name}{A short grid name used in directory and file name generation.}
}
\value{

}
\description{
This function extracts a set of point location forecasts from a previously
downloaded gridded NOAA Climate (Coupled) Forecast System (CFS) forecast and optionally
temporally downscales them.
}
