% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_noaa_historical.R
\name{stack_noaa_forecasts}
\alias{stack_noaa_forecasts}
\title{Function to stack together the first six hours of each NOAA forecast cycle and append them to create a stacked data product}
\usage{
stack_noaa_forecasts(
  forecast_dates,
  site,
  noaa_directory,
  noaa_model,
  output_directory,
  model_name = "observed-met-noaa",
  dates_w_errors = NA,
  s3_mode = FALSE,
  bucket = NULL,
  verbose = FALSE
)
}
\arguments{
\item{forecast_dates}{vector; vector of dates for which you have NOAA GEFS forecasts with all four cycles (00, 06, 12, 18)}

\item{site}{character; four letter site name}

\item{noaa_directory}{filepath;  directory where you have noaa forecasts stored}

\item{noaa_model}{character;  name of the noaa model you are using, e.g. "noaa/NOAAGEFS_6hr"}

\item{output_directory}{filepath; directory where the output files from this function will go}

\item{model_name}{character; name of the model output from this function}

\item{dates_w_errors}{list; list of dates that cause errors, e.g. have missing first timestep}
}
\value{
, returns a netCDF file for each NOAA GEFS ensemble (0-31) with the first six hours of each NOAA GEFS forecast cycle stacked to produce a continuous meteorological data product. Returns both 6hr files and temporally downscaled 1hr files for each ensembles
}
\description{
Function to stack together the first six hours of each NOAA forecast cycle and append them to create a stacked data product
}
