% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_noaa_gefs_netcdf.R
\name{write_noaa_gefs_netcdf}
\alias{write_noaa_gefs_netcdf}
\title{Write a (NOAA GEFS?) forecast to netCDF}
\usage{
write_noaa_gefs_netcdf(
  df,
  ens = NA,
  lat,
  lon,
  cf_units,
  output_file,
  overwrite
)
}
\arguments{
\item{df}{A data frame of meteorological variables to be written to a netCDF file. Columns
must start with 'time' with the following columns in the order of \code{cf_units}.}

\item{ens}{Optional: An ensemble index used for subsetting df.}

\item{lat}{Latitude for the forecast location in degree north.}

\item{lon}{Longitude for the forecast location in degree east.}

\item{cf_units}{Vector of variable unit descriptor strings in the order they appear in df.}

\item{output_file}{Name, with full path, of the netcdf file that is generated.}

\item{overwrite}{Logical stating whether to overwrite an existing output file.}
}
\value{
None
}
\description{
This function takes a data frame containing meteorological forecast data and
related metadata and saves it to a netCDF file.
}
\author{
Quinn Thomas
}
