% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_historical_stacked.R
\name{average_stacked_forecasts}
\alias{average_stacked_forecasts}
\title{Average ensemble from stacked historical noaa forecasts}
\usage{
average_stacked_forecasts(forecast_dates, noaa_stacked_directory, output_file)
}
\arguments{
\item{forecast_dates}{vector of dates will be included in stacked file}

\item{noaa_stacked_directory}{full path to directory with the 6hr stacked files for the site}

\item{output_file}{name (with full path) of averaged stack file}
}
\value{

}
\description{
Average ensemble from stacked historical noaa forecasts
}
