% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_cfs_grid_process_downscale.R
\name{noaa_cfs_grid_process_downscale}
\alias{noaa_cfs_grid_process_downscale}
\title{Extract and temporally downscale points from downloaded grid files}
\usage{
noaa_cfs_grid_process_downscale(
  lat_list,
  lon_list,
  site_list,
  downscale,
  debias = FALSE,
  overwrite,
  model_name,
  model_name_ds,
  model_name_ds_debias,
  model_name_raw,
  debias_coefficients = NULL,
  num_cores = 1,
  output_directory,
  reprocess = FALSE,
  grid_name
)
}
\arguments{
\item{output_directory}{}
}
\value{

}
\description{
Extract and temporally downscale points from downloaded grid files
}
\examples{

}
