% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_gefs_point_download_downscale.R
\name{noaa_gefs_point_download_downscale}
\alias{noaa_gefs_point_download_downscale}
\title{Download and Downscale NOAA GEFS for a single site}
\usage{
noaa_gefs_point_download_downscale(
  read_from_path,
  lat_list,
  lon_list,
  site_list,
  forecast_time = NA,
  forecast_date = NA,
  downscale,
  overwrite,
  model_name,
  model_name_ds,
  output_directory
)
}
\arguments{
\item{lat_list, }{vector of latitudes that correspond to site codes}

\item{lon_list, }{vector of longitudes that correspond to site codes}

\item{site_list, }{vector of site codes, used in directory and file name generation}

\item{downscale, }{logical specifying whether to downscale from 6-hr to 1-hr}

\item{overwrite, }{logical stating to overwrite any existing output_file}

\item{model_name, }{directory name for the 6-hr forecast, this will be used in directory and file name generation}

\item{model_name_ds, }{directory name for the 1-hr forecast, this will be used in directory and file name generation}

\item{output_directory, }{directory where the model output will be save}

\item{site_index, }{index of site_list, lat_list, lon_list to be downloaded}
}
\value{
None
}
\description{
Download and Downscale NOAA GEFS for a single site
}
\author{
Quinn Thomas
}
