% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_downscale.R
\name{temporal_downscale}
\alias{temporal_downscale}
\title{Downscale NOAA GEFS frin 6hr to 1hr}
\usage{
temporal_downscale(input_file, output_file, overwrite = TRUE)
}
\arguments{
\item{input_file, }{full path to 6hr file}

\item{output_file, }{full path to 1hr file that will be generated}

\item{overwrite, }{logical stating to overwrite any existing output_file}
}
\value{
None
}
\description{
Downscale NOAA GEFS frin 6hr to 1hr
}
\author{
Quinn Thomas
}
