% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_noaa_gefs_netcdf.R
\name{write_noaa_gefs_netcdf}
\alias{write_noaa_gefs_netcdf}
\title{Write NOAA GEGS netCDF}
\usage{
write_noaa_gefs_netcdf(df, ens, lat, lon, cf_units, output_file, overwrite)
}
\arguments{
\item{df, }{dataframe of data written to netCDF}

\item{VarNames, }{names of vars that are state variables}
}
\value{
NA
}
\description{
Write NOAA GEGS netCDF
}
\author{
Quinn Thomas
}
